+++ python/transitio/report/__init__.py
@@ -3,8 +3,17 @@
 from transitio.report._report import (
     build_report,
     parity_summary,
+    render_comparison_html,
+    render_comparison_markdown,
     render_html,
     render_markdown,
 )
 
-__all__ = ["build_report", "parity_summary", "render_html", "render_markdown"]
+__all__ = [
+    "build_report",
+    "parity_summary",
+    "render_comparison_html",
+    "render_comparison_markdown",
+    "render_html",
+    "render_markdown",
+]